backend/requirements.txt
tensorflow==2.0.0
google-cloud-storage==1.16.1
Pillow==6.0.0
numpy~=1.21.5
tensorflowjs